Split up report writer more
Preparing for project report generation.
