+++ ghostwriter/modules/reportwriter/report/json.py
@@ -1,0 +1,10 @@
+
+from ghostwriter.modules.reportwriter.base.json import ExportBaseJson
+from ghostwriter.modules.reportwriter.report.base import ExportReportBase
+
+
+class ExportReportJson(ExportBaseJson, ExportReportBase):
+    """
+    JSON exporter for reports
+    """
+    pass
