+++ ghostwriter/modules/reportwriter/richtext/ooxml.py
